+++ src/main/java/org/kumoricon/model/blacklist/BlacklistName.java
@@ -1,0 +1,52 @@
+package org.kumoricon.model.blacklist;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
+@Entity
+@Table(name = "blacklist")
+public class BlacklistName {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    @NotNull
+    private String firstName;
+    @NotNull
+    private String lastName;
+
+    public BlacklistName(String firstName, String lastName) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
+
+    public String getFirstName() { return firstName; }
+    public void setFirstName(String firstName) { this.firstName = firstName; }
+
+    public String getLastName() { return lastName; }
+    public void setLastName(String lastName) { this.lastName = lastName; }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof BlacklistName)) return false;
+
+        BlacklistName that = (BlacklistName) o;
+
+        if (!firstName.equals(that.firstName)) return false;
+        return lastName.equals(that.lastName);
+    }
+
+    @Override
+    public int hashCode() {
+        int result = firstName.hashCode();
+        result = 31 * result + lastName.hashCode();
+        return result;
+    }
+
+    public String toString() {
+        return String.format("[Blacklist %s: %s %s]", id, firstName, lastName);
+    }
+}